Move up communicating report section
Seems like people will not notice it if it's after a list of the ways to receive a report. People who are in the middle of receiving a report will want to see the "how to communicate a report" section immediately after looking to see if they've gathered the right set of information.