Replace "harasser" with "reported person"
There are types of Code of Conduct reports where a reported person isn't
a "harasser", or some people might argue that their behavior doesn't
meet a dictionary definition of harassment.

An example of this is microaggressions -- small acts that reinforce
structures of privilege and remind people from groups under-represented
at your event that they aren't welcome, acknowledged, or listened to. A
microagression could be calling an adult woman a "girl", not having an
accessible venue, assuming a racial minority is a food server rather
than an attendee, etc.

These small acts add up over time, contributing to people feeling
unwelcome at events. Someone may face these kinds of microaggressions
many times over the course of the day. These microaggressions are a Code
of Conduct violation, because they make a member of one of the protected
classes in your Code of Conduct feel unwelcome because of their
identity.

Not everyone who experiences microaggressions will classify it as
"harassment". It's important to use neutral language when describing the
person they reported, in order to allow for people who feel
uncomfortable and unwelcome because of someone's behavior, but wouldn't
describe it as "harassment".

People who perpetuate microaggressions may be unaware of what they are
doing, and tend to become defensive when the word "harasser" is used.
Other people are aware that what they are doing is inappropriate, and
they are harassers. Regardless of whether someone intended to be a
"harasser", their behavior makes attendees feel unsafe or unwelcome, and
the inappropriate behavior needs to stop.

There are also situations where behavior is inappropriate, but wouldn't
be classified as harassment. For example, two volunteers may end up
verbally abusing each other in an argument, or a talk reviewer might
take a bribe to upvote their friend's talk. Neither of those situations
are about harassment, but they are both inappropriate behavior that
doesn't follow the community guidelines.

Thus, refer to a "harasser" as "the reported person".